# 8484 Implement aggregate sum and use for arc counters