@tonyhutter 2017-07-12
What kind of performance increases are you seeing?