@prakashsurya 2017-07-12
@ahrens @pcd1193182 can either of you comment on the performance of this? from what I remember, it was a little complicated, due to some other bottlenecks that we haven't yet pushed fixes for.

looking at our internal review, it says this patch increased performance from ~9300 MB/s to ~11200 MB/s; but that was tested with some other experimental patches to address bottlenecks elsewhere on the system. without these additional patches, the performance improvement was modest, going from ~5810 MB/s to 5870 MB/s. with that said, I think the bottlenecks seen elsewhere might be illumos specific, so it'd be interesting to get results from another platform.