@idodeclare 2017-08-30
Hi, Prakash,

The former implementation of dbuf.c's dbuf_hash() kept the lower six bits of the the 64-bit value of (void *os) out of the hash calculation. Likewise, the former implementation of arc.c's buf_hash() kept the lower eight bits of (uint64_t spa) out of the hash calculation. Is there no meaningful impact to your new implementation using those formerly-excluded bits in the cityhash?

Regards.