@pcd1193182 2017-09-02
Hi @idodeclare, including the lower 8 bits of the spa and many more bits of the os (because we're dropping the lower 8 and then & with 0xFF, so we only get bits 8-15) does not decrease the quality of the hashing. It will not cause any new collisions which did not already occur, and we perform a full equality check to find the appropriate element, so it will not result in incorrect return values.