@idodeclare 2017-09-02
Hi, @pcd1193182. I was inquiring about the inverse—whether including more information in hash will cause hashing to diverge where it was previously equal for similar values of \`os' or \`spa'.