@pcd1193182 2017-09-03
That can only be a good thing, no? Fewer hash collisions will generally result in higher performance. Even if you do lots of lookups to similar-argument dbufs at about the same time, we store them using chaining, so it doesn't get you any cache effect wins.