@idodeclare 2017-09-05
To be clearer, mine was not a question of quality of hashing—and certainly not of more collisions—but of identity, which you answered with "perform a full equality check to find the appropriate element." That was not clear from the diff. I.e. revision of the included bits without comment was what spurred my question.