@pcd1193182 2017-09-05
@idodeclare Ah, I understand now. Sorry for the confusion. You can see the equality check in dbuf.c immediately below the diff, in dbuf_find, and in arc.c in buf_hash_find.