@prakashsurya 2018-01-10
There were some concerns about this patch raised during the RTI, see here for more details: https://illumos.topicbox.com/groups/advocates/Te0a7c923a738145f-M0fcb9388d77093699c17c54d

I'm closing this, and will re-open when I get some time to do the testing requested in the RTI thread linked above.